fix: Typescript build by adding static types dependency list
By default typescript tries to build any given types available in its build scope but that brings some issue when the consumer project import types dependencies because tsc would then try to compile them as well whereas they should remains out of the build scope of the library.

In order to solve this I've statically listed all types dependencies of this project in the types attribute of ts config file.